fix: handle errors gracefully

+++ lib/devise/rownd/strategies/rownd_authenticatable.rb
@@ -20,18 +20,26 @@ module Devise
         @access_token = session[:rownd_user_access_token]
         return fail!('No Access Token') unless @access_token
 
-        @decoded_jwt = verify_token(@access_token)
-        @app_id = @decoded_jwt['aud'].find(/^app:.+/).first.split(':').last
+        begin
+          @decoded_jwt = verify_token(@access_token)
 
-        configured_app_id = Devise::Rownd.app_id
-        ok = @app_id == configured_app_id
-        return fail!('JWT not authorized for app') unless ok
+          @app_id = @decoded_jwt['aud'].find(/^app:.+/).first.split(':').last
 
-        rownd_user = Devise::Rownd::User.new(fetch_user)
+          configured_app_id = Devise::Rownd.app_id
+          ok = @app_id == configured_app_id
+          return fail!('JWT not authorized for app') unless ok
 
-        return fail!(:unable_to_authenticate) unless rownd_user
+          user_data = fetch_user
+          return fail!('Failed to fetch user') unless user_data
 
-        success!(rownd_user)
+          rownd_user = Devise::Rownd::User.new(user_data)
+
+          return fail!('Failed to initialize user') unless rownd_user
+
+          success!(rownd_user)
+        rescue StandardError => e
+          fail!("Unable to authenticate: #{e.message}")
+        end
       end
 
       def return_to_after_sign_out
@@ -42,13 +50,18 @@ module Devise
         cache_key = "rownd_user_#{@decoded_jwt['jti']}"
         if session[:rownd_stale_data] == true
           data = fetch_user_from_api
+          return nil unless data
+
           Rails.cache.write(cache_key, data, expires_in: 1.minute)
           session.delete(:rownd_stale_data) if session[:rownd_stale_data]
           return data
         end
 
         Rails.cache.fetch(cache_key, expires_in: 1.minute) do
-          fetch_user_from_api
+          fetched_user = fetch_user_from_api
+          break unless fetched_user
+
+          fetched_user
         end
       end
 
@@ -62,31 +75,37 @@ module Devise
         )
         return response.body['data'] if response.success?
 
-        raise StandardError, response.body['message']
+        Rails.logger.error("Failed to fetch user: #{response.body['message']}")
+        nil
       end
 
       def verify_token(access_token)
-        for jwk in jwks
-          begin
-            response = JOSE::JWT.verify_strict(jwk, ['EdDSA'], access_token)
-            return response[1].fields if response[0]
-          rescue StandardError => e
-            puts "Error: #{e}"
-            next
-          end
-          raise StandardError
+        raise StandardError, 'No JWKs' unless jwks
+
+        jwks.each do |jwk|
+          response = JOSE::JWT.verify_strict(jwk, ['EdDSA'], access_token)
+          return response[1].fields if response[0]
+        rescue StandardError
+          next
         end
+        raise StandardError, 'Failed to verify JWT. No matching JWKs'
       end
 
       def jwks
         Rails.cache.fetch('rownd_jwks', expires_in: 15.minutes) do
-          fetch_jwks
+          fetched_jwks = fetch_jwks_from_api
+          break unless fetched_jwks
+
+          fetched_jwks
         end
       end
 
-      def fetch_jwks
+      def fetch_jwks_from_api
         response = ::Devise::Rownd::API.make_api_call('/hub/auth/keys')
-        response.body['keys']
+        return response.body['keys'] if response.success?
+
+        Rails.logger.error("Failed to fetch JWKs: #{response.body['message']}")
+        nil
       end
     end
   end
